BOT 3.2.0/requirements.txt
discord
autopy
opencv-python
numpy
pillow
pywin32
pyautogui
keyboard
pynput
pywinauto
pygetwindow
python-vlc